Use `:normal!` to avoid running afoul of remapped '@'
This is a potential issue reported here:

https://github.com/wincent/replay/issues/4

For consistency, I changed the other `feedkeys(..., 'n')` to `:normal!`
as well, even though the "n" flag there already projects us from
unintended remapping. There is another `feedkeys()` higher up in the
file but I left that as is because it is probably tidier than doing
`:execute "normal! ..."` etc.

Closes: https://github.com/wincent/replay/issues/4